Made a few changes to the scripts. See extended description for full details.
* Changed some integers to enums, to facilitate selections when using
the Inspector.
* Added GameConstants.cs, so as to conveniently stores constants that
are used in many scripts in a single place.
* Implemented Whole and Half rests, which transport non-Whole followers
up and down. Both Whole and Half rests are under the "Rest" prefab, so
you have to select what type of rest it is once you place it in the
scene.
* Added three different types of Followers in FollowerMove (Whole
notes, Quarter notes, and Eighth notes):
   - Whole notes are slower, and stop rests from warping other
followers up and down.
   - Quarter notes are your good ol' followers. Same mechanics as
before.
   - Eighth notes can 'jump' along with the player, regardless on if
the player is touching it at the moment or not.
* Note that you can set the note_type for each follower in the
inspector using a drop-down list.